Move tests in test_common.py to test_helpers.py
